move more enums into constants
